20180326 v2 征免性质 支持代码搜索

+++ custom_addons/customs_center/models/customs_declaration.py
@@ -948,11 +948,12 @@ class CustomsDeclaration(models.Model):
     #                 _logger.info(
     #                     u'Had parsed the attach xml message %s' % xml_attach_message.decode('utf-8'))
 
+
     # @api.model
     # @q_job.job
     @api.multi
     def auto_parse_attach_message_xml(self):
-        """ 自动解析随附单据入库 从随附单据报文到报关单 """
+        """ 自动解析随附单据入库 从随附单据报文到报关单 反向查找"""
         # company_xml_parse_path = '0000016165'  # 做成前端界面可配置
 
         # 先判断随附单据队列模型里是否有数据
@@ -979,15 +980,17 @@ class CustomsDeclaration(models.Model):
 
         # 首先解析随附单据目录的文件 可能多个附件
         attach_files = os.listdir(parse_attach_path)
-        attach_files_list = {attach_filename.split('$')[0]: attach_filename for attach_filename in attach_files if
-                             attach_filename.endswith('.xml')}
+        attach_files_list = {attach_filename.split('$')[0] : attach_filename for attach_filename in attach_files if attach_filename.endswith('.xml')}
+
 
         if not attach_files_list:
             return True
         # attach_files = {os.path.join(parse_attach_path, i) for i in attach_files_list}
 
         # 读文件，用lxml解析报文
+        attach_ids = []  # 附件id用于将解析的随附单据加在 随附单据拖拽上传page页
         attach_name_list = []
+        # edoc_queue_ids = edoc_queue_ids.ids
         for edoc_queue_obj in edoc_queue_ids:
             name = edoc_queue_obj.edoc_id
             datas_fname = edoc_queue_obj.edoc_id
@@ -999,8 +1002,7 @@ class CustomsDeclaration(models.Model):
             xml_attach_message = os.path.join(parse_attach_path, attach_files_list[name])
             with open(xml_attach_message, 'r') as f:
                 attach_xml_str = str(f.read())
-                attach_xml_str1 = attach_xml_str.replace('xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"',
-                                                         '')
+                attach_xml_str1 = attach_xml_str.replace('xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"', '')
                 attach_xml_str = attach_xml_str1.replace('xsi:nil="true"', '')
                 # print xml_str
                 root = etree.fromstring(attach_xml_str)  # 打开xml文档
@@ -1012,7 +1014,7 @@ class CustomsDeclaration(models.Model):
                     attach_data_node = root.xpath('.//TcsData')
                     for child in attach_data_node[0]:
                         xml_attach_message_dic[child.tag] = child.text
-                attach_name_in_xml = xml_attach_message_dic.get('FILE_NAME')  # 获取随附单据报文中的文件名
+                attach_name_in_xml = xml_attach_message_dic.get('FILE_NAME')   # 获取随附单据报文中的文件名
                 binary_data = xml_attach_message_dic.get('BINARY_DATA', None)  # 获取随附单据报文中的二进制数据
 
                 # # 根据上述获取的附件名称 在附件模型中查找 对应的附件ID
@@ -1025,17 +1027,25 @@ class CustomsDeclaration(models.Model):
                 # print(res_id)
 
                 # 根据队列里的内容创建附件， 并删除相应的记录
-                self.env['ir.attachment'].create({
-                    'name': name,
-                    'datas_fname': datas_fname,
-                    'extension': 'pdf',
-                    'res_model': 'customs_center.customs_dec',
-                    'res_id': cus_dec_id,
-                    'dec_edoc_type': dec_edoc_type,
-                    'datas': binary_data
-                })
-                edoc_queue_obj.unlink()
-                attach_name_list.append(xml_attach_message)
+                if attach_name_in_xml == name:
+                    attach_file_obj = self.env['ir.attachment'].create({
+                        'name': name,
+                        'datas_fname': datas_fname,
+                        'extension': 'pdf',
+                        'res_model': 'customs_center.customs_dec',
+                        'res_id': cus_dec_id,
+                        'dec_edoc_type': dec_edoc_type,
+                        'datas': binary_data
+                    })
+
+                    edoc_queue_obj.unlink()
+                    attach_name_list.append(xml_attach_message)
+
+                    attach_ids.append(attach_file_obj.id)
+                    # 附件id用于将解析的随附单据加在 随附单据拖拽上传page页
+                    customs_dec_model_obj = self.env['customs_center.customs_dec'].search([('id', '=', cus_dec_id)])
+                    customs_dec_model_obj.information_attachment_ids = [(6, 0, attach_ids)]
+
 
                 # # 根据上方找到的报关单ID 找到该报关单对应的附件列表
                 # information_attachment_ids = self.env['ir.attachment'].search(
@@ -1052,10 +1062,9 @@ class CustomsDeclaration(models.Model):
 
         # 将解析成功的随附单据报文 移动到随附单据备份目录
         for xml_attach_message in attach_name_list:  # xml_attach_message是单据名
-            shutil.move(xml_attach_message, backup_attach_xml_path)
-            _logger.info(
-                u'Had parsed the attach xml message %s' % xml_attach_message.decode('utf-8'))
-
+                shutil.move(xml_attach_message, backup_attach_xml_path)
+                _logger.info(
+                    u'Had parsed the attach xml message %s' % xml_attach_message.decode('utf-8'))
 
     @api.multi
     def generate_single_customer_xml_after(self):
